@@ -53,22 +53,21 @@
     ]
   },
   "dependencies": {
-    "@sentry/react": "^6.17.4",
+    "@sentry/react": "^6.19.7",
     "@walletconnect/web3-provider": "^1.8.0",
     "classnames": "^2.3.1",
     "date-fns": "^2.28.0",
     "ethers": "^5.5.4",
     "immer": "^9.0.12",
     "lodash": "^4.17.21",
-    "rc-tooltip": "^5.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-focus-lock": "^2.7.1",
-    "react-helmet-async": "^1.2.2",
+    "rc-tooltip": "^5.2.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-focus-lock": "^2.9.2",
+    "react-helmet-async": "^1.3.0",
     "react-input-autosize": "^3.0.0",
-    "react-number-format": "^4.9.1",
-    "react-responsive-carousel": "^3.2.22",
-    "react-router-dom": "^5.2.0",
+    "react-number-format": "^4.9.4",
+    "react-router-dom": "^5.3.4",
     "react-toastify": "^7.0.4",
     "web3modal": "^1.9.5"
   },
@@ -79,18 +78,18 @@
     "@openzeppelin/contracts": "4.2.0",
     "@percy/cli": "^1.6.0",
     "@percy/cypress": "^3.1.2",
-    "@testing-library/cypress": "^8.0.3",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/cypress": "^8.0.7",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "@typechain/ethers-v5": "^8.0.2",
     "@types/chai": "^4.3.0",
     "@types/jest": "^27.5.2",
     "@types/lodash": "^4.14.178",
     "@types/mocha": "^9.1.0",
     "@types/node": "^17.0.15",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "@types/react-input-autosize": "^2.2.1",
     "@types/react-router-dom": "^5.3.3",
     "api3-dao": "git+https://github.com/api3dao/api3-dao.git#0b8bb17fbda0161d5d9fff00ed4562b91cc27af6",
@@ -123,6 +122,7 @@
   },
   "resolutions": {
     "@types/jest": "^27.5.2",
+    "@types/react": "^18.0.26",
     "jest-regex-util": "^27.5.1",
     "jest-watcher": "^27.5.1",
     "jest-worker": "^27.5.1"
