@@ -1,5 +1,4 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import * as Sentry from '@sentry/react';
 import './index.scss';
 import App from './app';
@@ -30,9 +29,9 @@ if (process.env.REACT_APP_NODE_ENV === 'development' && (window as any).ethereum
   mockLocalhostWeb3Provider(window);
 }
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const root = createRoot(document.getElementById('root')!);
+
+// NOTE: Strict mode triggers useEffect and useMemo hooks twice on mount (in development), which results
+// in our data hooks triggering 3 times on mount (the third trigger is caused by the memoized smart contracts
+// which are included in the useEffect dependency arrays). Because of this, we have decided to turn off strict mode.
+root.render(<App />);
